refactor: base root entity 분리
