+++ backend/src/main/java/com/yigongil/backend/domain/certification/Certification.java
@@ -1,6 +1,6 @@
 package com.yigongil.backend.domain.certification;
 
-import com.yigongil.backend.domain.BaseEntity;
+import com.yigongil.backend.domain.base.BaseEntity;
 import com.yigongil.backend.domain.member.Member;
 import com.yigongil.backend.domain.round.Round;
 import com.yigongil.backend.domain.study.Study;
